Make SD card configurable via webserver

+++ Software/src/devboard/webserver/debug_logging_html.cpp
@@ -57,9 +57,9 @@ String debug_logger_processor(void) {
     content += "<button onclick='refreshPage()'>Refresh data</button> ";
   }
   content += "<button onclick='exportLog()'>Export to .txt</button> ";
-#ifdef LOG_TO_SD
-  content += "<button onclick='deleteLog()'>Delete log file</button> ";
-#endif
+  if (datalayer.system.info.SD_logging_active) {
+    content += "<button onclick='deleteLog()'>Delete log file</button> ";
+  }
   content += "<button onclick='goToMainPage()'>Back to main page</button>";
 
   // Start a new block for the debug log messages
@@ -98,9 +98,9 @@ String debug_logger_processor(void) {
   content += "<script>";
   content += "function refreshPage(){ location.reload(true); }";
   content += "function exportLog() { window.location.href = '/export_log'; }";
-#ifdef LOG_TO_SD
-  content += "function deleteLog() { window.location.href = '/delete_log'; }";
-#endif
+  if (datalayer.system.info.SD_logging_active) {
+    content += "function deleteLog() { window.location.href = '/delete_log'; }";
+  }
   content += "function goToMainPage() { window.location.href = '/'; }";
   content += "</script>";
   content += index_html_footer;
